Add authentication to delete recipe with tests

+++ web/src/Application/Controller/DeleteRecipe.php
@@ -6,6 +6,8 @@ namespace HelloFresh\Application\Controller;
 use League\Tactician\CommandBus;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use HelloFresh\Domain\Command\DeleteRecipe as DeleteRecipeCommand;
 
 final class DeleteRecipe
@@ -13,13 +15,21 @@ final class DeleteRecipe
     /** @var CommandBus */
     private $commandBus;
 
-    public function __construct(CommandBus $commandBus)
+    /** @var TokenStorageInterface */
+    private $tokenStorage;
+
+    public function __construct(CommandBus $commandBus, TokenStorageInterface $tokenStorage)
     {
         $this->commandBus = $commandBus;
+        $this->tokenStorage = $tokenStorage;
     }
 
     public function __invoke(Request $request): JsonResponse
     {
+        if (!in_array('ROLE_ADMIN', $this->tokenStorage->getToken()->getRoleNames())) {
+            throw new UnauthorizedHttpException('Basic');
+        }
+
         $command = new DeleteRecipeCommand(
             $request->get('id')
         );
